add deriveKeys to autoConnect flow
