+++ src/views/dialogs/OnboardingDialog/GenerateKeys.tsx
@@ -1,20 +1,13 @@
-import { useState } from 'react';
-
-import { AES } from 'crypto-js';
 import styled, { css } from 'styled-components';
 
 import { EvmDerivedAccountStatus } from '@/constants/account';
 import { AlertType } from '@/constants/alerts';
-import { AnalyticsEvents } from '@/constants/analytics';
 import { ButtonAction } from '@/constants/buttons';
 import { STRING_KEYS } from '@/constants/localization';
-import { DydxAddress, WalletType } from '@/constants/wallets';
+import { WalletType } from '@/constants/wallets';
 
+import { useGenerateKeys } from '@/hooks/Onboarding/useGenerateKeys';
 import { useAccounts } from '@/hooks/useAccounts';
-import { useDydxClient } from '@/hooks/useDydxClient';
-import { useEnvConfig } from '@/hooks/useEnvConfig';
-import { useMatchingEvmNetwork } from '@/hooks/useMatchingEvmNetwork';
-import useSignForWalletDerivation from '@/hooks/useSignForWalletDerivation';
 import { useSimpleUiEnabled } from '@/hooks/useSimpleUiEnabled';
 import { useStringGetter } from '@/hooks/useStringGetter';
 
@@ -26,13 +19,7 @@ import { GreenCheckCircle } from '@/components/GreenCheckCircle';
 import { LoadingSpinner } from '@/components/Loading/LoadingSpinner';
 import { WithReceipt } from '@/components/WithReceipt';
 
-import { useAppDispatch } from '@/state/appTypes';
-import { setSavedEncryptedSignature } from '@/state/wallet';
-
-import { track } from '@/lib/analytics/analytics';
 import { isTruthy } from '@/lib/isTruthy';
-import { log } from '@/lib/telemetry';
-import { parseWalletError } from '@/lib/wallet';
 
 type ElementProps = {
   status: EvmDerivedAccountStatus;
@@ -42,155 +29,21 @@ type ElementProps = {
 
 export const GenerateKeys = ({ status, setStatus, onKeysDerived = () => {} }: ElementProps) => {
   const stringGetter = useStringGetter();
-  const dispatch = useAppDispatch();
-  const { sourceAccount, setWalletFromSignature } = useAccounts();
   const isSimpleUi = useSimpleUiEnabled();
+  const { sourceAccount } = useAccounts();
 
-  const [error, setError] = useState<string>();
-
-  // 1. Switch network
-  const ethereumChainId = useEnvConfig('ethereumChainId');
-  const chainId = Number(ethereumChainId);
-
-  const { isMatchingNetwork, matchNetwork, isSwitchingNetwork } = useMatchingEvmNetwork({
-    chainId,
+  const {
+    error,
+    isDeriving,
+    isMatchingNetwork,
+    isSwitchingNetwork,
+    onClickSwitchNetwork,
+    onClickSendRequestOrTryAgain,
+  } = useGenerateKeys({
+    status,
+    setStatus,
+    onKeysDerived,
   });
-
-  const switchNetwork = async () => {
-    setError(undefined);
-
-    try {
-      await matchNetwork();
-      return true;
-    } catch (err) {
-      const { message, walletErrorType, isErrorExpected } = parseWalletError({
-        error: err,
-        stringGetter,
-      });
-
-      if (!isErrorExpected) {
-        log('GenerateKeys/switchNetwork', err, { walletErrorType });
-      }
-
-      if (message) {
-        setError(message);
-      }
-
-      return false;
-    }
-  };
-
-  const switchNetworkAndDeriveKeys = async () => {
-    const networkSwitched = await switchNetwork();
-    if (networkSwitched) await deriveKeys();
-  };
-
-  // 2. Derive keys from EVM account
-  const { getWalletFromSignature } = useDydxClient();
-  const { getSubaccounts } = useAccounts();
-
-  const isDeriving = ![
-    EvmDerivedAccountStatus.NotDerived,
-    EvmDerivedAccountStatus.Derived,
-  ].includes(status);
-
-  const signMessageAsync = useSignForWalletDerivation(sourceAccount.walletInfo);
-
-  const staticEncryptionKey = import.meta.env.VITE_PK_ENCRYPTION_KEY;
-
-  const deriveKeys = async () => {
-    setError(undefined);
-
-    try {
-      // 1. First signature
-      setStatus(EvmDerivedAccountStatus.Deriving);
-
-      const signature = await signMessageAsync();
-      track(
-        AnalyticsEvents.OnboardingDeriveKeysSignatureReceived({
-          signatureNumber: 1,
-        })
-      );
-      const { wallet: dydxWallet } = await getWalletFromSignature({ signature });
-
-      // 2. Ensure signature is deterministic
-      // Check if subaccounts exist
-      const dydxAddress = dydxWallet.address as DydxAddress;
-      let hasPreviousTransactions = false;
-
-      try {
-        const subaccounts = await getSubaccounts({ dydxAddress });
-        hasPreviousTransactions = subaccounts.length > 0;
-
-        track(AnalyticsEvents.OnboardingAccountDerived({ hasPreviousTransactions }));
-
-        if (!hasPreviousTransactions) {
-          setStatus(EvmDerivedAccountStatus.EnsuringDeterminism);
-
-          // Second signature
-          const additionalSignature = await signMessageAsync();
-          track(
-            AnalyticsEvents.OnboardingDeriveKeysSignatureReceived({
-              signatureNumber: 2,
-            })
-          );
-
-          if (signature !== additionalSignature) {
-            throw new Error(
-              'Your wallet does not support deterministic signing. Please switch to a different wallet provider.'
-            );
-          }
-        }
-      } catch (err) {
-        setStatus(EvmDerivedAccountStatus.NotDerived);
-        const { message } = parseWalletError({ error: err, stringGetter });
-
-        if (message) {
-          track(AnalyticsEvents.OnboardingWalletIsNonDeterministic());
-          setError(message);
-        }
-        return;
-      }
-
-      await setWalletFromSignature(signature);
-
-      // 3: Remember me (encrypt and store signature)
-      if (staticEncryptionKey) {
-        const encryptedSignature = AES.encrypt(signature, staticEncryptionKey).toString();
-        dispatch(setSavedEncryptedSignature(encryptedSignature));
-      }
-
-      // 4. Done
-      setStatus(EvmDerivedAccountStatus.Derived);
-    } catch (err) {
-      setStatus(EvmDerivedAccountStatus.NotDerived);
-      const { message, walletErrorType, isErrorExpected } = parseWalletError({
-        error: err,
-        stringGetter,
-      });
-
-      if (message) {
-        setError(message);
-        if (!isErrorExpected) {
-          log('GenerateKeys/deriveKeys', err, { walletErrorType });
-        }
-      }
-    }
-  };
-
-  const onClickSwitchNetwork = () => {
-    switchNetworkAndDeriveKeys().then(onKeysDerived);
-    track(AnalyticsEvents.OnboardingSwitchNetworkClick());
-  };
-
-  const onClickSendRequestOrTryAgain = () => {
-    deriveKeys().then(onKeysDerived);
-    track(
-      AnalyticsEvents.OnboardingSendRequestClick({
-        firstAttempt: !error,
-      })
-    );
-  };
 
   return (
     <>
@@ -307,6 +160,7 @@ export const GenerateKeys = ({ status, setStatus, onKeysDerived = () => {} }: El
     </>
   );
 };
+
 const $StatusCard = styled.div<{ active?: boolean }>`
   ${layoutMixins.spacedRow}
   gap: 1rem;
